requirements.txt
Django==2.0.3
django-recaptcha2==1.0.3
django-summernote==0.8.8.6
django-treebeard==4.2.1
docutils==0.14
Pillow==5.0.0
PyJWT==1.6.1
python-dateutil==2.7.0
pytz==2018.3
sqlparse==0.2.4
django-allauth==0.37.1
django-ordered-model==2.1.0
